cli/operator: reorganize into focused files, finalize naming, address review
Squashes the post-refactor reorganization, review fixes, and polish that sit on
top of the atomic refactor commits:

- Final vocabulary: run()->runNode(), assemble()->newNode(), the assembled
  struct->node.
- Split the catch-all files into focused ones by concern - keys.go, db.go,
  network.go, eventsync.go, and start_node.go (the cobra command + bootstrap) -
  leaving node.go as node construction + lifecycle and config.go as config +
  validation.
- Cleanup-contract fixes: close the db on migration failure; close the EL client
  in node.Close() and on newNode() failure; make the empty execution-address
  check effective; clearer key-mismatch error messages.
- Conventions/polish: fmt.Errorf over errors.New; tighter doc comments.
- Tests mirror their source files (config/network/node _test.go), uniformly
  named Test_<func>.

Gate green (gofmt/build/vet/test -race/lint); the file and identifier moves are
relocation-only.
