--- cli/operator/assemble_test.go
@@ -1,206 +1,0 @@
-package operator
-
-import (
-	"context"
-	"fmt"
-	"testing"
-
-	api "github.com/attestantio/go-eth2-client/api"
-	eth2apiv1 "github.com/attestantio/go-eth2-client/api/v1"
-	"github.com/stretchr/testify/require"
-	"go.uber.org/zap"
-
-	"github.com/ssvlabs/ssv/doppelganger"
-	"github.com/ssvlabs/ssv/eth/executionclient"
-	"github.com/ssvlabs/ssv/exporter"
-	"github.com/ssvlabs/ssv/hprobe"
-	"github.com/ssvlabs/ssv/network"
-	"github.com/ssvlabs/ssv/networkconfig"
-	"github.com/ssvlabs/ssv/ssvsigner/keys"
-)
-
-// stubBeaconClient satisfies the in-package beaconClient interface by embedding it (nil): it
-// promotes every method so the type compiles as a beaconClient, while implementing only what the
-// test path actually invokes. In operator mode with Doppelganger disabled, assemble() never calls
-// a beacon method synchronously (it only stores the client); the one async caller is the
-// NewVRSubmitter goroutine, which the test stops by canceling ctx before its first slot tick.
-type stubBeaconClient struct {
-	beaconClient
-}
-
-// SetProposalPreparationsProvider is the one beacon method invoked synchronously during assembly
-// (operator.New wires it into the fee-recipient controller); a no-op satisfies the smoke test.
-func (stubBeaconClient) SetProposalPreparationsProvider(func() ([]*eth2apiv1.ProposalPreparation, error)) {
-}
-
-// SubmitValidatorRegistrations is called unconditionally on each slot tick by the NewVRSubmitter
-// goroutine assemble() starts in operator mode. A no-op makes the smoke test structurally safe if
-// that goroutine ticks before ctx is canceled, rather than relying on the slot interval winning the
-// race (it would otherwise hit the embedded-nil beacon and panic).
-func (stubBeaconClient) SubmitValidatorRegistrations(context.Context, []*api.VersionedSignedValidatorRegistration) error {
-	return nil
-}
-
-// stubExecutionClient satisfies executionclient.Provider the same way. assemble() only stores the
-// EL client (it is consumed in start(), not here), so no method is invoked during assembly.
-type stubExecutionClient struct {
-	executionclient.Provider
-}
-
-// Test_assemble_wiresOperatorNode is the in-process smoke test for the assemble() seam: with
-// stubbed beacon/EL clients and a minimal operator-mode config, the full wiring graph
-// (db → storage → keys → p2p → validator controller → operator node) must construct without
-// error. It exercises everything up to — but not including — start()'s network bring-up, which
-// performs real socket I/O. The doppelganger dimension covers buildDoppelganger's real-handler
-// path (enabled) alongside the no-op path (disabled).
-func Test_assemble_wiresOperatorNode(t *testing.T) {
-	for _, tc := range []struct {
-		name           string
-		doppelgangerOn bool
-	}{
-		{name: "doppelganger disabled", doppelgangerOn: false},
-		{name: "doppelganger enabled", doppelgangerOn: true},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			logger := zap.NewNop()
-
-			operatorPrivKey, err := keys.GeneratePrivateKey()
-			require.NoError(t, err)
-
-			netCfg := *networkconfig.TestNetwork
-			networkConfig := &netCfg
-
-			cfg := &config{}
-			cfg.OperatorPrivateKey = operatorPrivKey.Base64()
-			cfg.DBOptions.Path = t.TempDir()
-			// Keep all long-lived servers off; they only matter in start().
-			cfg.MetricsAPIPort = 0
-			cfg.SSVAPIPort = 0
-			cfg.WsAPIPort = 0
-			cfg.EnableDoppelgangerProtection = tc.doppelgangerOn
-
-			res := resolved{mode: modeOperator, usingPrivKey: true}
-
-			a, err := assemble(ctx, cfg, logger, res, networkConfig, stubBeaconClient{}, stubExecutionClient{})
-			require.NoError(t, err)
-			require.NotNil(t, a)
-			require.NotNil(t, a.db)
-			require.NotNil(t, a.nodeStorage)
-			require.NotNil(t, a.p2pNetwork)
-			require.NotNil(t, a.validatorCtrl)
-			require.NotNil(t, a.operatorNode)
-
-			// buildDoppelganger returns the no-op handler when protection is disabled and a real
-			// handler when enabled.
-			_, isNoOp := a.doppelgangerHandler.(doppelganger.NoOpHandler)
-			require.Equal(t, !tc.doppelgangerOn, isNoOp, "doppelganger handler must match the configured protection")
-
-			// Stop the VRSubmitter goroutine (started in assemble) before tearing down, then ensure
-			// the CLI-owned resources close cleanly — the p2p network was constructed but never
-			// Setup/Start'd.
-			cancel()
-			require.NoError(t, a.Close())
-		})
-	}
-}
-
-// Test_assemble_wiresExporterNode mirrors the operator smoke test for the exporter paths: with no
-// signing identity, it asserts assemble() wires the graph for both exporter modes and that the
-// mode-specific divergences hold — no key manager in either, and a duty-trace collector only in
-// archive mode.
-func Test_assemble_wiresExporterNode(t *testing.T) {
-	for _, tc := range []struct {
-		name          string
-		mode          nodeMode
-		exporterMode  string
-		wantCollector bool
-	}{
-		{name: "standard", mode: modeExporterStandard, exporterMode: exporter.ModeStandard, wantCollector: false},
-		{name: "archive", mode: modeExporterArchive, exporterMode: exporter.ModeArchive, wantCollector: true},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			ctx, cancel := context.WithCancel(context.Background())
-			defer cancel()
-
-			logger := zap.NewNop()
-
-			netCfg := *networkconfig.TestNetwork
-			networkConfig := &netCfg
-
-			cfg := &config{}
-			cfg.DBOptions.Path = t.TempDir()
-			cfg.ExporterOptions.Enabled = true
-			cfg.ExporterOptions.Mode = tc.exporterMode
-			cfg.MetricsAPIPort = 0
-			cfg.SSVAPIPort = 0
-			cfg.WsAPIPort = 0
-
-			// Exporter mode resolves no signing flags; mode alone drives the divergence.
-			res := resolved{mode: tc.mode}
-
-			a, err := assemble(ctx, cfg, logger, res, networkConfig, stubBeaconClient{}, stubExecutionClient{})
-			require.NoError(t, err)
-			require.NotNil(t, a)
-			require.NotNil(t, a.db)
-			require.NotNil(t, a.nodeStorage)
-			require.NotNil(t, a.p2pNetwork)
-			require.NotNil(t, a.validatorCtrl)
-			require.NotNil(t, a.operatorNode)
-			require.Nil(t, a.keyManager, "exporter nodes have no key manager")
-
-			if tc.wantCollector {
-				require.NotNil(t, a.collector, "archive mode wires a duty-trace collector")
-			} else {
-				require.Nil(t, a.collector, "standard mode has no duty-trace collector")
-			}
-
-			cancel()
-			require.NoError(t, a.Close())
-		})
-	}
-}
-
-// recordingP2PNetwork is a stub network.P2PNetwork that records Setup/Start instead of performing
-// real socket I/O, and satisfies p2pv1.HealthChecker (via Healthy) so startNetwork's health-prober
-// registration — a.p2pNetwork.(p2pv1.HealthChecker) — succeeds.
-type recordingP2PNetwork struct {
-	network.P2PNetwork
-	setupCalled bool
-	startCalled bool
-}
-
-func (r *recordingP2PNetwork) Setup() error                  { r.setupCalled = true; return nil }
-func (r *recordingP2PNetwork) Start() error                  { r.startCalled = true; return nil }
-func (r *recordingP2PNetwork) Healthy(context.Context) error { return nil }
-
-// Test_assembled_startNetwork_wiresStatsRegardlessOfDynamicMaxPeers is the CLI-level regression for
-// the DynamicMaxPeers bug: GetValidatorStats must be wired and the p2p network Set up + Started
-// whether or not DynamicMaxPeers is enabled. Re-nesting these under the DynamicMaxPeers flag (the
-// original bug) would fail this test.
-func Test_assembled_startNetwork_wiresStatsRegardlessOfDynamicMaxPeers(t *testing.T) {
-	for _, dynamicMaxPeers := range []bool{false, true} {
-		t.Run(fmt.Sprintf("DynamicMaxPeers=%t", dynamicMaxPeers), func(t *testing.T) {
-			stubNet := &recordingP2PNetwork{}
-			cfg := &config{}
-			cfg.P2pNetworkConfig.DynamicMaxPeers = dynamicMaxPeers
-
-			a := &assembled{
-				logger:     zap.NewNop(),
-				cfg:        cfg,
-				p2pNetwork: stubNet,
-				// validatorCtrl stays nil: startNetwork only *assigns* the GetValidatorStats closure
-				// (pubsub invokes it later, not here), so a nil controller is never dereferenced.
-			}
-
-			require.NoError(t, a.startNetwork(hprobe.NewHealthProber(zap.NewNop())))
-
-			require.NotNil(t, cfg.P2pNetworkConfig.GetValidatorStats,
-				"GetValidatorStats must be wired regardless of DynamicMaxPeers")
-			require.True(t, stubNet.setupCalled, "p2p Setup must run regardless of DynamicMaxPeers")
-			require.True(t, stubNet.startCalled, "p2p Start must run regardless of DynamicMaxPeers")
-		})
-	}
-}
