+++ cli/operator/config_test.go
@@ -1,7 +1,9 @@
 package operator
 
 import (
-	"errors"
+	"fmt"
+	"net"
+	"strconv"
 	"testing"
 	"time"
 
@@ -11,6 +13,10 @@ import (
 	"go.uber.org/zap/zaptest/observer"
 
 	"github.com/ssvlabs/ssv/exporter"
+	"github.com/ssvlabs/ssv/networkconfig"
+	operatorstorage "github.com/ssvlabs/ssv/operator/storage"
+	kv "github.com/ssvlabs/ssv/storage/badger"
+	"github.com/ssvlabs/ssv/storage/basedb"
 )
 
 const (
@@ -22,42 +28,18 @@ const (
 	msgInvalidExporterMode = "invalid exporter mode"
 )
 
-func TestValidateProposerDelay(t *testing.T) {
-	tests := []struct {
-		name           string
-		delay          time.Duration
-		allowDangerous bool
-		wantErr        bool
-	}{
-		{name: "zero -> ok", delay: 0, allowDangerous: false, wantErr: false},
-		{name: "100ms -> ok", delay: 100 * time.Millisecond, allowDangerous: false, wantErr: false},
-		{name: "300ms (recommended start) -> ok", delay: 300 * time.Millisecond, allowDangerous: false, wantErr: false},
-		{name: "at limit 1000ms -> ok", delay: 1000 * time.Millisecond, allowDangerous: false, wantErr: false},
-		{name: "1001ms without flag -> error", delay: 1001 * time.Millisecond, allowDangerous: false, wantErr: true},
-		{name: "2000ms without flag -> error", delay: 2000 * time.Millisecond, allowDangerous: false, wantErr: true},
-		{name: "5000ms without flag -> error", delay: 5000 * time.Millisecond, allowDangerous: false, wantErr: true},
-		{name: "1001ms with flag -> ok", delay: 1001 * time.Millisecond, allowDangerous: true, wantErr: false},
-		{name: "5000ms with flag -> ok", delay: 5000 * time.Millisecond, allowDangerous: true, wantErr: false},
-	}
+func Test_config_load(t *testing.T) {
+	var c config
+	require.NoError(t, c.load("", "")) // both paths unset -> no-op
 
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := validateProposerDelay(tt.delay, tt.allowDangerous)
-			if tt.wantErr {
-				require.Error(t, err)
-				require.Contains(t, err.Error(), "ProposerDelay value")
-				require.Contains(t, err.Error(), "exceeds maximum safe delay")
-				require.Contains(t, err.Error(), "AllowDangerousProposerDelay")
-				require.Contains(t, err.Error(), "ALLOW_DANGEROUS_PROPOSER_DELAY")
-				return
-			}
-			require.NoError(t, err)
-		})
-	}
+	require.ErrorContains(t, c.load("/nonexistent/config.yaml", ""),
+		"could not read config needed for logger initialization")
+	require.ErrorContains(t, c.load("", "/nonexistent/share.yaml"),
+		"could not read share config needed for logger initialization")
 }
 
 // Test_resolveAndValidate_proposerDelay covers the proposer-delay advisory warning emitted by
-// resolveAndValidate (validation itself is covered by TestValidateProposerDelay). Signing is
+// resolveAndValidate (validation itself is covered by Test_validateProposerDelay). Signing is
 // left unset + non-exporter so resolveSigning is a no-op for these cases.
 func Test_resolveAndValidate_proposerDelay(t *testing.T) {
 	t.Run("dangerous delay with flag - warns with duration fields", func(t *testing.T) {
@@ -110,8 +92,280 @@ func Test_resolveAndValidate_proposerDelay(t *testing.T) {
 	})
 }
 
-// Test_resolveSigning covers signing-method resolution + mutual-exclusivity validation. On
-// stage these conflicts were logger.Fatal calls (untested); they are now returned errors.
+// Test_resolveAndValidate_signingErrorContext verifies resolveAndValidate enriches a signing
+// error with the configured-source context, without exposing the private key value.
+func Test_resolveAndValidate_signingErrorContext(t *testing.T) {
+	c := config{}
+	c.SSVSigner.Endpoint = testSignerEndpoint
+	c.OperatorPrivateKey = testOperatorKey
+
+	_, err := c.resolveAndValidate(zap.NewNop())
+	require.Error(t, err)
+	require.Contains(t, err.Error(), "cannot enable both remote signing")
+
+	// start_node.go logs runNode()'s error via startupErrorLogFields — the signing-source context
+	// must be preserved as queryable structured fields, without exposing the key value.
+	core, recorded := observer.New(zapcore.ErrorLevel)
+	zap.New(core).Error("could not start node", startupErrorLogFields(err)...)
+	require.Len(t, recorded.All(), 1)
+
+	m := recorded.All()[0].ContextMap()
+	require.Equal(t, testSignerEndpoint, m["ssv_signer_endpoint"])
+	require.Contains(t, m, "private_key_file")
+	require.Contains(t, m, "password_file")
+	require.EqualValues(t, len(testOperatorKey), m["operator_private_key_len"])
+	for _, v := range m {
+		require.NotEqual(t, testOperatorKey, v) // the private key value itself is never logged
+	}
+}
+
+// Test_resolveAndValidate_mode verifies the operating mode is resolved into the result and that
+// an invalid EXPORTER_MODE fails validation up front.
+func Test_resolveAndValidate_mode(t *testing.T) {
+	t.Run("non-exporter -> modeOperator", func(t *testing.T) {
+		c := config{}
+		res, err := c.resolveAndValidate(zap.NewNop())
+		require.NoError(t, err)
+		require.Equal(t, modeOperator, res.mode)
+	})
+
+	t.Run("exporter archive -> modeExporterArchive", func(t *testing.T) {
+		c := config{}
+		c.ExporterOptions.Enabled = true
+		c.ExporterOptions.Mode = exporter.ModeArchive
+		res, err := c.resolveAndValidate(zap.NewNop())
+		require.NoError(t, err)
+		require.Equal(t, modeExporterArchive, res.mode)
+	})
+
+	t.Run("invalid exporter mode -> error", func(t *testing.T) {
+		c := config{}
+		c.ExporterOptions.Enabled = true
+		c.ExporterOptions.Mode = "bogus"
+		_, err := c.resolveAndValidate(zap.NewNop())
+		require.Error(t, err)
+		require.Contains(t, err.Error(), msgInvalidExporterMode)
+	})
+}
+
+func Test_validateProposerDelay(t *testing.T) {
+	tests := []struct {
+		name           string
+		delay          time.Duration
+		allowDangerous bool
+		wantErr        bool
+	}{
+		{name: "zero -> ok", delay: 0, allowDangerous: false, wantErr: false},
+		{name: "100ms -> ok", delay: 100 * time.Millisecond, allowDangerous: false, wantErr: false},
+		{name: "300ms (recommended start) -> ok", delay: 300 * time.Millisecond, allowDangerous: false, wantErr: false},
+		{name: "at limit 1000ms -> ok", delay: 1000 * time.Millisecond, allowDangerous: false, wantErr: false},
+		{name: "1001ms without flag -> error", delay: 1001 * time.Millisecond, allowDangerous: false, wantErr: true},
+		{name: "2000ms without flag -> error", delay: 2000 * time.Millisecond, allowDangerous: false, wantErr: true},
+		{name: "5000ms without flag -> error", delay: 5000 * time.Millisecond, allowDangerous: false, wantErr: true},
+		{name: "1001ms with flag -> ok", delay: 1001 * time.Millisecond, allowDangerous: true, wantErr: false},
+		{name: "5000ms with flag -> ok", delay: 5000 * time.Millisecond, allowDangerous: true, wantErr: false},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			err := validateProposerDelay(tt.delay, tt.allowDangerous)
+			if tt.wantErr {
+				require.Error(t, err)
+				require.Contains(t, err.Error(), "ProposerDelay value")
+				require.Contains(t, err.Error(), "exceeds maximum safe delay")
+				require.Contains(t, err.Error(), "AllowDangerousProposerDelay")
+				require.Contains(t, err.Error(), "ALLOW_DANGEROUS_PROPOSER_DELAY")
+				return
+			}
+			require.NoError(t, err)
+		})
+	}
+}
+
+func Test_validateConfig(t *testing.T) {
+	logger := zap.New(zapcore.NewNopCore(), zap.WithFatalHook(zapcore.WriteThenPanic))
+
+	db, err := kv.NewInMemory(logger, basedb.Options{})
+	require.NoError(t, err)
+
+	netCfg := networkconfig.TestNetwork
+	nodeStorage, err := operatorstorage.NewNodeStorage(netCfg.Beacon, logger, db)
+	require.NoError(t, err)
+
+	testNetworkName := netCfg.StorageName()
+
+	t.Run("no config in DB", func(t *testing.T) {
+		c := &operatorstorage.ConfigLock{
+			NetworkName:      testNetworkName,
+			UsingLocalEvents: true,
+			UsingSSVSigner:   true,
+		}
+		require.NoError(t, validateConfig(nodeStorage, c.NetworkName, c.UsingLocalEvents, c.UsingSSVSigner, false))
+
+		storedConfig, found, err := nodeStorage.GetConfig(nil)
+		require.NoError(t, err)
+		require.True(t, found)
+		require.Equal(t, c, storedConfig)
+
+		require.NoError(t, nodeStorage.DeleteConfig(nil))
+	})
+
+	t.Run("has same config in DB", func(t *testing.T) {
+		c := &operatorstorage.ConfigLock{
+			NetworkName:      testNetworkName,
+			UsingLocalEvents: true,
+			UsingSSVSigner:   true,
+		}
+		require.NoError(t, nodeStorage.SaveConfig(nil, c))
+		require.NoError(t, validateConfig(nodeStorage, c.NetworkName, c.UsingLocalEvents, c.UsingSSVSigner, false))
+
+		storedConfig, found, err := nodeStorage.GetConfig(nil)
+		require.NoError(t, err)
+		require.True(t, found)
+		require.Equal(t, c, storedConfig)
+
+		require.NoError(t, nodeStorage.DeleteConfig(nil))
+	})
+
+	t.Run("has different network name, events type, and ssv signer in DB", func(t *testing.T) {
+		c := &operatorstorage.ConfigLock{
+			NetworkName:      testNetworkName + "1",
+			UsingLocalEvents: false,
+			UsingSSVSigner:   false,
+		}
+		require.NoError(t, nodeStorage.SaveConfig(nil, c))
+		require.ErrorContains(t,
+			validateConfig(nodeStorage, testNetworkName, true, true, false),
+			"incompatible config change: network mismatch. Stored network testnet:alan1 does not match current network testnet:alan. The database must be removed or reinitialized",
+		)
+
+		storedConfig, found, err := nodeStorage.GetConfig(nil)
+		require.NoError(t, err)
+		require.True(t, found)
+		require.Equal(t, c, storedConfig)
+
+		require.NoError(t, nodeStorage.DeleteConfig(nil))
+	})
+
+	t.Run("has different network name in DB", func(t *testing.T) {
+		c := &operatorstorage.ConfigLock{
+			NetworkName:      testNetworkName + "1",
+			UsingLocalEvents: true,
+			UsingSSVSigner:   true,
+		}
+		require.NoError(t, nodeStorage.SaveConfig(nil, c))
+		require.ErrorContains(t,
+			validateConfig(nodeStorage, testNetworkName, c.UsingLocalEvents, c.UsingSSVSigner, false),
+			"incompatible config change: network mismatch. Stored network testnet:alan1 does not match current network testnet:alan. The database must be removed or reinitialized",
+		)
+
+		storedConfig, found, err := nodeStorage.GetConfig(nil)
+		require.NoError(t, err)
+		require.True(t, found)
+		require.Equal(t, c, storedConfig)
+
+		require.NoError(t, nodeStorage.DeleteConfig(nil))
+	})
+
+	t.Run("has real events in DB but runs with local events", func(t *testing.T) {
+		c := &operatorstorage.ConfigLock{
+			NetworkName:      testNetworkName,
+			UsingLocalEvents: false,
+			UsingSSVSigner:   true,
+		}
+		require.NoError(t, nodeStorage.SaveConfig(nil, c))
+		require.ErrorContains(t,
+			validateConfig(nodeStorage, c.NetworkName, true, true, false),
+			"incompatible config change: enabling local events is not allowed. The database must be removed or reinitialized",
+		)
+
+		storedConfig, found, err := nodeStorage.GetConfig(nil)
+		require.NoError(t, err)
+		require.True(t, found)
+		require.Equal(t, c, storedConfig)
+
+		require.NoError(t, nodeStorage.DeleteConfig(nil))
+	})
+
+	t.Run("has local events in DB but runs with real events", func(t *testing.T) {
+		c := &operatorstorage.ConfigLock{
+			NetworkName:      testNetworkName,
+			UsingLocalEvents: true,
+			UsingSSVSigner:   true,
+		}
+		require.NoError(t, nodeStorage.SaveConfig(nil, c))
+		require.ErrorContains(t,
+			validateConfig(nodeStorage, c.NetworkName, false, true, false),
+			"incompatible config change: disabling local events is not allowed. The database must be removed or reinitialized",
+		)
+
+		storedConfig, found, err := nodeStorage.GetConfig(nil)
+		require.NoError(t, err)
+		require.True(t, found)
+		require.Equal(t, c, storedConfig)
+
+		require.NoError(t, nodeStorage.DeleteConfig(nil))
+	})
+
+	t.Run("has local signer in DB but runs with remote signer", func(t *testing.T) {
+		c := &operatorstorage.ConfigLock{
+			NetworkName:      testNetworkName,
+			UsingLocalEvents: true,
+			UsingSSVSigner:   true,
+		}
+		require.NoError(t, nodeStorage.SaveConfig(nil, c))
+		require.ErrorContains(t,
+			validateConfig(nodeStorage, c.NetworkName, true, false, false),
+			"incompatible config change: disabling ssv-signer is not allowed. The database must be removed or reinitialized",
+		)
+
+		storedConfig, found, err := nodeStorage.GetConfig(nil)
+		require.NoError(t, err)
+		require.True(t, found)
+		require.Equal(t, c, storedConfig)
+
+		require.NoError(t, nodeStorage.DeleteConfig(nil))
+	})
+
+	t.Run("has remote signer in DB but runs with local signer", func(t *testing.T) {
+		c := &operatorstorage.ConfigLock{
+			NetworkName:      testNetworkName,
+			UsingLocalEvents: true,
+			UsingSSVSigner:   false,
+		}
+		require.NoError(t, nodeStorage.SaveConfig(nil, c))
+		require.ErrorContains(t,
+			validateConfig(nodeStorage, c.NetworkName, true, true, false),
+			"incompatible config change: enabling ssv-signer is not allowed. The database must be removed or reinitialized",
+		)
+
+		storedConfig, found, err := nodeStorage.GetConfig(nil)
+		require.NoError(t, err)
+		require.True(t, found)
+		require.Equal(t, c, storedConfig)
+
+		require.NoError(t, nodeStorage.DeleteConfig(nil))
+	})
+
+	t.Run("exporter ignores stored signer mode", func(t *testing.T) {
+		c := &operatorstorage.ConfigLock{
+			NetworkName:      testNetworkName,
+			UsingLocalEvents: true,
+			UsingSSVSigner:   true,
+		}
+		require.NoError(t, nodeStorage.SaveConfig(nil, c))
+		require.NoError(t, validateConfig(nodeStorage, c.NetworkName, true, false, true))
+
+		storedConfig, found, err := nodeStorage.GetConfig(nil)
+		require.NoError(t, err)
+		require.True(t, found)
+		require.Equal(t, c, storedConfig)
+
+		require.NoError(t, nodeStorage.DeleteConfig(nil))
+	})
+}
+
+// Test_resolveSigning covers signing-method resolution + mutual-exclusivity validation.
 func Test_resolveSigning(t *testing.T) {
 	tests := []struct {
 		name    string
@@ -167,63 +421,8 @@ func Test_resolveSigning(t *testing.T) {
 	}
 }
 
-// Test_resolveAndValidate_signingErrorContext verifies resolveAndValidate enriches a signing
-// error with the configured-source context (restored from the pre-refactor structured log
-// fields), without exposing the private key value.
-func Test_resolveAndValidate_signingErrorContext(t *testing.T) {
-	c := config{}
-	c.SSVSigner.Endpoint = testSignerEndpoint
-	c.OperatorPrivateKey = testOperatorKey
-
-	_, err := c.resolveAndValidate(zap.NewNop())
-	require.Error(t, err)
-	require.Contains(t, err.Error(), "cannot enable both remote signing")
-
-	// node.go logs run()'s error via startupErrorLogFields — the configured signing-source context
-	// must be preserved as queryable structured fields, without exposing the key value.
-	core, recorded := observer.New(zapcore.ErrorLevel)
-	zap.New(core).Error("could not start node", startupErrorLogFields(err)...)
-	require.Len(t, recorded.All(), 1)
-
-	m := recorded.All()[0].ContextMap()
-	require.Equal(t, testSignerEndpoint, m["ssv_signer_endpoint"])
-	require.Contains(t, m, "private_key_file")
-	require.Contains(t, m, "password_file")
-	require.EqualValues(t, len(testOperatorKey), m["operator_private_key_len"])
-	for _, v := range m {
-		require.NotEqual(t, testOperatorKey, v) // the private key value itself is never logged
-	}
-}
-
-// Test_startupErrorLogFields verifies the consolidated startup fatal preserves the structured
-// fields carried by a startupError (e.g. the ssv-signer endpoint), and degrades to just the
-// error for a plain error.
-func Test_startupErrorLogFields(t *testing.T) {
-	t.Run("startupError fields are attached", func(t *testing.T) {
-		err := startupError{
-			err:    errors.New("ssv-signer unavailable"),
-			fields: []zap.Field{zap.String("ssv_signer_endpoint", testSignerEndpoint)},
-		}
-		core, recorded := observer.New(zapcore.ErrorLevel)
-		zap.New(core).Error("could not start node", startupErrorLogFields(err)...)
-
-		m := recorded.All()[0].ContextMap()
-		require.Equal(t, testSignerEndpoint, m["ssv_signer_endpoint"])
-		require.Contains(t, m, "error")
-	})
-
-	t.Run("plain error -> no extra fields", func(t *testing.T) {
-		core, recorded := observer.New(zapcore.ErrorLevel)
-		zap.New(core).Error("could not start node", startupErrorLogFields(errors.New("boom"))...)
-
-		m := recorded.All()[0].ContextMap()
-		require.Contains(t, m, "error")
-		require.NotContains(t, m, "ssv_signer_endpoint")
-	})
-}
-
 // Test_resolveMode covers operating-mode resolution and the fail-fast rejection of an
-// unrecognized EXPORTER_MODE (previously a late Fatal in the exporter collector switch).
+// unrecognized EXPORTER_MODE.
 func Test_resolveMode(t *testing.T) {
 	tests := []struct {
 		name    string
@@ -254,42 +453,93 @@ func Test_resolveMode(t *testing.T) {
 	}
 }
 
-// Test_resolveAndValidate_mode verifies the operating mode is resolved into the result and that
-// an invalid EXPORTER_MODE fails validation up front (the rejection moved out of the late
-// exporter collector switch).
-func Test_resolveAndValidate_mode(t *testing.T) {
-	t.Run("non-exporter -> modeOperator", func(t *testing.T) {
-		c := config{}
-		res, err := c.resolveAndValidate(zap.NewNop())
-		require.NoError(t, err)
-		require.Equal(t, modeOperator, res.mode)
+func Test_warnIfSSVAPIAddressUnset(t *testing.T) {
+	t.Parallel()
+
+	t.Run("warns when address is empty", func(t *testing.T) {
+		core, recorded := observer.New(zapcore.WarnLevel)
+		logger := zap.New(core)
+
+		warnIfSSVAPIAddressUnset(logger, "", 16000)
+
+		logs := recorded.All()
+		require.Len(t, logs, 1)
+		require.Equal(t, zapcore.WarnLevel, logs[0].Level)
+		require.Equal(t, "SSV API address not configured; listening on all interfaces", logs[0].Message)
+		require.EqualValues(t, 16000, logs[0].ContextMap()["port"])
+		require.Equal(t, "SSVAPIAddress", logs[0].ContextMap()["config_key"])
+		require.Equal(t, "127.0.0.1", logs[0].ContextMap()["recommended_address"])
 	})
 
-	t.Run("exporter archive -> modeExporterArchive", func(t *testing.T) {
-		c := config{}
-		c.ExporterOptions.Enabled = true
-		c.ExporterOptions.Mode = exporter.ModeArchive
-		res, err := c.resolveAndValidate(zap.NewNop())
-		require.NoError(t, err)
-		require.Equal(t, modeExporterArchive, res.mode)
-	})
+	t.Run("does not warn when address is set", func(t *testing.T) {
+		core, recorded := observer.New(zapcore.WarnLevel)
+		logger := zap.New(core)
 
-	t.Run("invalid exporter mode -> error", func(t *testing.T) {
-		c := config{}
-		c.ExporterOptions.Enabled = true
-		c.ExporterOptions.Mode = "bogus"
-		_, err := c.resolveAndValidate(zap.NewNop())
-		require.Error(t, err)
-		require.Contains(t, err.Error(), msgInvalidExporterMode)
+		warnIfSSVAPIAddressUnset(logger, "127.0.0.1", 16000)
+
+		require.Len(t, recorded.All(), 0)
 	})
 }
 
-func Test_config_load(t *testing.T) {
-	var c config
-	require.NoError(t, c.load("", "")) // both paths unset -> no-op
+func Test_ssvAPIListenAddress(t *testing.T) {
+	t.Parallel()
 
-	require.ErrorContains(t, c.load("/nonexistent/config.yaml", ""),
-		"could not read config needed for logger initialization")
-	require.ErrorContains(t, c.load("", "/nonexistent/share.yaml"),
-		"could not read share config needed for logger initialization")
+	testCases := []struct {
+		name    string
+		address string
+		port    int
+		want    string
+	}{
+		{
+			name: "empty address listens on all interfaces",
+			port: 16000,
+			want: ":16000",
+		},
+		{
+			name:    "ipv4 loopback",
+			address: "127.0.0.1",
+			port:    16000,
+			want:    "127.0.0.1:16000",
+		},
+		{
+			name:    "ipv6 loopback",
+			address: "::1",
+			port:    16000,
+			want:    "[::1]:16000",
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			got := net.JoinHostPort(tc.address, strconv.Itoa(tc.port))
+			require.Equal(t, tc.want, got)
+		})
+	}
+}
+
+// Test_startupErrorLogFields verifies the consolidated startup fatal preserves the structured
+// fields carried by a startupError (e.g. the ssv-signer endpoint), and degrades to just the
+// error for a plain error.
+func Test_startupErrorLogFields(t *testing.T) {
+	t.Run("startupError fields are attached", func(t *testing.T) {
+		err := startupError{
+			err:    fmt.Errorf("ssv-signer unavailable"),
+			fields: []zap.Field{zap.String("ssv_signer_endpoint", testSignerEndpoint)},
+		}
+		core, recorded := observer.New(zapcore.ErrorLevel)
+		zap.New(core).Error("could not start node", startupErrorLogFields(err)...)
+
+		m := recorded.All()[0].ContextMap()
+		require.Equal(t, testSignerEndpoint, m["ssv_signer_endpoint"])
+		require.Contains(t, m, "error")
+	})
+
+	t.Run("plain error -> no extra fields", func(t *testing.T) {
+		core, recorded := observer.New(zapcore.ErrorLevel)
+		zap.New(core).Error("could not start node", startupErrorLogFields(fmt.Errorf("boom"))...)
+
+		m := recorded.All()[0].ContextMap()
+		require.Contains(t, m, "error")
+		require.NotContains(t, m, "ssv_signer_endpoint")
+	})
 }
